Allow for gl views of sizes different than their source UIViews
We want to allow UIFXViews to be larger than the UIViews which they
receive their snapshots from so that we can account for a model which
stretches beyond its original bounds. On the venetian effect, the
corners of the strips extend beyond the edges of the source view during
the animation.