Set light position after setting the view matrix
We weren't getting any changes to the shadow darkness when the blind was
rotating. This was do to the fact that we were setting the light
position after setting the modelViewMaxtrix meaning that we were
effectivly rotating the light source with each strip so the shadow never
changed darkness. Now we have a method that sets the viewMatrix (which
represents the camera position only), right at the start of the render
cycle. This is where we set the light position. The light position is
affected by the current modelViewMaxtrix of the GLKBaseEffect so we want
so set the modelViewMaxtrix of GLKBaseEffect to the camera position only
and not change it after.